app/portfolio/requirements.txt
Django==2.2.13
uwsgi==2.0.18
mysqlclient==1.4.6
django-bootstrap4==1.1.1
django-cleanup==3.2.0
django-filter==2.2.0
djangorestframework==3.11.0
django-webtest==1.9.7
pillow==7.0.0
requests==2.23.0
Faker==0.8.16
factory-boy==2.11.1
pytest==5.4.3
django-environ==0.4.5
boto3==1.14.24
django-storages==1.9.1